requirements.txt
insightface==0.7.3
onnxruntime-gpu==1.19.2
opencv-python>=4.8,<5
tqdm>=4.66,<5
numpy>=1.23,<2
